# openapi.json
{"openapi":"3.1.0","info":{"title":"Supernova","version":"0.1.0"},"paths":{"/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_register_auth_register_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/logout/{login}":{"post":{"tags":["auth"],"summary":"Logout","operationId":"logout_auth_logout__login__post","security":[{"LoginManager":[]}],"parameters":[{"name":"login","in":"path","required":true,"schema":{"type":"string","title":"Login"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/ping":{"get":{"tags":["auth"],"summary":"Auth Ping","operationId":"auth_ping_auth_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPing"}}}}},"security":[{"LoginManager":[]}]}},"/watchlist/":{"get":{"tags":["watchlist"],"summary":"Get Watchlist","operationId":"get_watchlist_watchlist__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Film"},"type":"array","title":"Response Get Watchlist Watchlist  Get"}}}}},"security":[{"LoginManager":[]}]}},"/watchlist/add":{"post":{"tags":["watchlist"],"summary":"Add New Film To Watchlist","operationId":"add_new_film_to_watchlist_watchlist_add_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilmAdd"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"LoginManager":[]}]}},"/watchlist/add-with-image":{"post":{"tags":["watchlist"],"summary":"Add New Film With Image To Watchlist","operationId":"add_new_film_with_image_to_watchlist_watchlist_add_with_image_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_add_new_film_with_image_to_watchlist_watchlist_add_with_image_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"LoginManager":[]}]}},"/watchlist/add/{film_id}":{"post":{"tags":["watchlist"],"summary":"Add Existing Film To Watchlist","operationId":"add_existing_film_to_watchlist_watchlist_add__film_id__post","security":[{"LoginManager":[]}],"parameters":[{"name":"film_id","in":"path","required":true,"schema":{"type":"integer","title":"Film Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/watchlist/remove/{film_id}":{"delete":{"tags":["watchlist"],"summary":"Remove Film From Watchlist","operationId":"remove_film_from_watchlist_watchlist_remove__film_id__delete","security":[{"LoginManager":[]}],"parameters":[{"name":"film_id","in":"path","required":true,"schema":{"type":"integer","title":"Film Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/films/add-image":{"post":{"tags":["films"],"summary":"Upload Film Image","operationId":"upload_film_image_films_add_image_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_film_image_films_add_image_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Image"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"LoginManager":[]}]}},"/films/discuss/{film_name}/{year}":{"get":{"tags":["films"],"summary":"Get Themes For Discussion","operationId":"get_themes_for_discussion_films_discuss__film_name___year__get","security":[{"LoginManager":[]}],"parameters":[{"name":"film_name","in":"path","required":true,"schema":{"type":"string","title":"Film Name"}},{"name":"year","in":"path","required":true,"schema":{"type":"integer","title":"Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"},"title":"Response Get Themes For Discussion Films Discuss  Film Name   Year  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/films/discuss-id/{film_id}":{"get":{"tags":["films"],"summary":"Get Themes For Discussion By Id","operationId":"get_themes_for_discussion_by_id_films_discuss_id__film_id__get","security":[{"LoginManager":[]}],"parameters":[{"name":"film_id","in":"path","required":true,"schema":{"type":"integer","title":"Film Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"},"title":"Response Get Themes For Discussion By Id Films Discuss Id  Film Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/films/":{"get":{"tags":["films"],"summary":"Get Films","operationId":"get_films_films__get","security":[{"LoginManager":[]}],"parameters":[{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}},{"name":"genres","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"type":"string"}},{"type":"null"}],"title":"Genres"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FilmReturn"},"title":"Response Get Films Films  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/films/{film_id}":{"get":{"tags":["films"],"summary":"Get Film","operationId":"get_film_films__film_id__get","security":[{"LoginManager":[]}],"parameters":[{"name":"film_id","in":"path","required":true,"schema":{"type":"integer","title":"Film Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilmReturn"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/watched/":{"get":{"tags":["watched"],"summary":"Get Watched","operationId":"get_watched_watched__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Film"},"type":"array","title":"Response Get Watched Watched  Get"}}}}},"security":[{"LoginManager":[]}]}},"/watched/add":{"post":{"tags":["watched"],"summary":"Add New Film To Watched","operationId":"add_new_film_to_watched_watched_add_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilmAdd"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"LoginManager":[]}]}},"/watched/add-with-image":{"post":{"tags":["watched"],"summary":"Add New Film With Image To Watched","operationId":"add_new_film_with_image_to_watched_watched_add_with_image_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_add_new_film_with_image_to_watched_watched_add_with_image_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"LoginManager":[]}]}},"/watched/add/{film_id}":{"post":{"tags":["watched"],"summary":"Add Existing Film To Watched","operationId":"add_existing_film_to_watched_watched_add__film_id__post","security":[{"LoginManager":[]}],"parameters":[{"name":"film_id","in":"path","required":true,"schema":{"type":"integer","title":"Film Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/session/create/{user_login}":{"post":{"tags":["session"],"summary":"Create Session","operationId":"create_session_session_create__user_login__post","security":[{"LoginManager":[]}],"parameters":[{"name":"user_login","in":"path","required":true,"schema":{"type":"string","title":"User Login"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Genres"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FilmReturn"},"title":"Response Create Session Session Create  User Login  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/session/end/{user_login}":{"post":{"tags":["session"],"summary":"End Session","operationId":"end_session_session_end__user_login__post","security":[{"LoginManager":[]}],"parameters":[{"name":"user_login","in":"path","required":true,"schema":{"type":"string","title":"User Login"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/profile/":{"get":{"tags":["profile"],"summary":"Get Profile","operationId":"get_profile_profile__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Profile"}}}}},"security":[{"LoginManager":[]}]}},"/":{"get":{"summary":"Ping","operationId":"ping__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}}},"components":{"schemas":{"Body_add_new_film_with_image_to_watched_watched_add_with_image_post":{"properties":{"film":{"type":"string","title":"Film"},"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["film","file"],"title":"Body_add_new_film_with_image_to_watched_watched_add_with_image_post"},"Body_add_new_film_with_image_to_watchlist_watchlist_add_with_image_post":{"properties":{"film":{"type":"string","title":"Film"},"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["film","file"],"title":"Body_add_new_film_with_image_to_watchlist_watchlist_add_with_image_post"},"Body_login_auth_login_post":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"Body_login_auth_login_post"},"Body_register_auth_register_post":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"Body_register_auth_register_post"},"Body_upload_film_image_films_add_image_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_film_image_films_add_image_post"},"Film":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"genres":{"items":{"type":"string"},"type":"array","title":"Genres"},"year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating"},"film_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Film Url"}},"type":"object","required":["id","title","rating"],"title":"Film"},"FilmAdd":{"properties":{"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"genres":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Genres"},"year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating"},"film_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Film Url"}},"type":"object","required":["title"],"title":"FilmAdd"},"FilmReturn":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"genres":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Genres","default":[]},"year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating"},"film_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Film Url"},"is_watchlisted":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Watchlisted","default":false}},"type":"object","required":["rating"],"title":"FilmReturn"},"Genres":{"properties":{"genres":{"items":{"type":"string"},"type":"array","title":"Genres"}},"type":"object","required":["genres"],"title":"Genres"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Image":{"properties":{"image_url":{"type":"string","title":"Image Url"}},"type":"object","required":["image_url"],"title":"Image"},"MessageResponse":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"MessageResponse"},"Profile":{"properties":{"username":{"type":"string","title":"Username"},"watched_count":{"type":"integer","title":"Watched Count"},"watchlist_count":{"type":"integer","title":"Watchlist Count"},"watched_films":{"items":{"$ref":"#/components/schemas/Film"},"type":"array","title":"Watched Films"}},"type":"object","required":["username","watched_count","watchlist_count","watched_films"],"title":"Profile"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"}},"type":"object","required":["access_token"],"title":"Token"},"UserPing":{"properties":{"username":{"type":"string","title":"Username"}},"type":"object","required":["username"],"title":"UserPing"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"LoginManager":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}